ibor-ai-gateway/pyproject.toml
[project]
name = "ai-gateway"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "kartikm76@yahoo.com", email = "kartikm76@yahoo.com" }
]
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.0.0",
    "openai-agents>=0.6.1",
    "fastapi>=0.119.1",
    "httpx>=0.28.1",
    "pydantic>=2.12.3",
    "pyyaml>=6.0",
    "python-dotenv>=1.1.1",
    "uvicorn[standard]>=0.38.0",
    "psycopg[binary]>=3.2",
    "psycopg-pool>=3.2",
    "pytest>=8.3.0",
    "yfinance>=0.2.50",
    "sentence-transformers>=3.0.0",
]

[project.scripts]
ai-gateway = "ai_gateway.main:app"

[build-system]
requires = ["uv_build>=0.9.5,<0.10.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
pythonpath = ["src"]
